+++ packages/documentation/src/pages/examples/DefaultItemRenderer.tsx
@@ -3,37 +3,44 @@ import { makeStyles } from '@material-ui/styles';
 import { Theme } from '@material-ui/core';
 import { Flipped } from 'react-flip-toolkit';
 
-import { ItemRendererProps } from 'react-sortly/src';
+import { ItemRendererProps, useDrag, useDrop, useIsClosestDragging } from 'react-sortly/src';
 
 type ItemItemRendererProps = ItemRendererProps<{
   name: string;
 }>;
-
-const useStyles = makeStyles((theme: Theme) => ({
-  root: (props: ItemItemRendererProps) => ({
+const useStyles = makeStyles<
+Theme, { muted: boolean; depth: number }>((theme: Theme) => ({
+  root: (props) => ({
     position: 'relative',
     marginBottom: theme.spacing(1.5),
-    zIndex: props.isDragging || props.isClosestDragging() ? 1 : 0,
+    zIndex: props.muted ? 1 : 0,
   }),
-  body: (props: ItemItemRendererProps) => ({
+  body: (props) => ({
     background: '#fff',
     cursor: 'move',
     padding: theme.spacing(2),
-    marginLeft: theme.spacing(props.data.depth * 2),
-    boxShadow: props.isDragging || props.isClosestDragging() ? '0px 0px 8px #666' : '0px 0px 2px #666',
-    border: props.isDragging || props.isClosestDragging() ? '1px dashed #1976d2' : '1px solid transparent',
+    marginLeft: theme.spacing(props.depth * 2),
+    boxShadow: props.muted ? '0px 0px 8px #666' : '0px 0px 2px #666',
+    border: props.muted ? '1px dashed #1976d2' : '1px solid transparent',
   }),
 }));
 
 const DefaultItemRenderer = (props: ItemItemRendererProps) => {
-  const { data: { id, name }, drag, drop } = props;
-  const ref = React.useRef< HTMLDivElement>(null);
-  const classes = useStyles(props);
-  drag(drop(ref));
+  const { id, depth, data: { name } } = props;
+  const [{ isDragging }, drag] = useDrag({
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging(),
+    }),
+  });
+  const [, drop] = useDrop();
+  const classes = useStyles({
+    muted: useIsClosestDragging() || isDragging,
+    depth,
+  });
 
   return (
     <Flipped flipId={id}>
-      <div ref={ref} className={classes.root}>
+      <div ref={(ref) => drag(drop(ref))} className={classes.root}>
         <div className={classes.body}>{name}</div>
       </div>
     </Flipped>