+++ packages/documentation/src/pages/examples/TableSortable/ItemRenderer.tsx
@@ -1,36 +1,41 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/styles';
-import { IconButton, TableRow, TableCell } from '@material-ui/core';
+import { Theme, IconButton, TableRow, TableCell } from '@material-ui/core';
 import ReorderIcon from '@material-ui/icons/Reorder';
 import { Flipped } from 'react-flip-toolkit';
 
-import { ItemRendererProps } from 'react-sortly/src';
+import { ItemRendererProps, useDrag, useDrop } from 'react-sortly/src';
 
 type ItemItemRendererProps = ItemRendererProps<{
   name: string;
 }>;
 
-const useStyles = makeStyles({
-  root: (props: ItemItemRendererProps) => ({
-    boxShadow: props.isDragging ? '0px 0px 8px #666' : 'none',
-    pointerEvents: props.isDragging ? 'none' : 'initial',
+const useStyles = makeStyles<
+Theme, { muted: boolean; depth: number }>({
+  root: (props) => ({
+    boxShadow: props.muted ? '0px 0px 8px #666' : 'none',
+    pointerEvents: props.muted ? 'none' : 'initial',
   }),
 });
 
 const ItemRenderer = (props: ItemItemRendererProps) => {
-  const { id, data: { name }, drag, drop, preview } = props;
-  const dropRef = React.useRef<any>(null);
-  const moveHandlerRef = React.useRef<HTMLButtonElement | null>(null);
-  const classes = useStyles(props);
-  
-  drag(moveHandlerRef);
-  drop(preview(dropRef));
+  const { id, depth, data: { name } } = props;
+  const [{ isDragging }, drag, preview] = useDrag({
+    collect: (monitor) => ({
+      isDragging: monitor.isDragging(),
+    }),
+  });
+  const [, drop] = useDrop();
 
+  const classes = useStyles({
+    muted: isDragging,
+    depth,
+  });
   return (
     <Flipped flipId={id}>
-      <TableRow ref={dropRef} className={classes.root}>
+      <TableRow ref={(ref: HTMLTableRowElement) => drop(preview(ref))} className={classes.root}>
         <TableCell style={{ width: 40 }}>
-          <IconButton ref={moveHandlerRef}><ReorderIcon /></IconButton>
+          <IconButton ref={drag}><ReorderIcon /></IconButton>
         </TableCell>
         <TableCell>
           {name}