+++ packages/documentation/src/routes.ts
@@ -1,4 +1,3 @@
-import Component from './pages/api/Component';
 import Simple from './pages/examples/Simple';
 import Advanced from './pages/examples/Advanced/Advanced';
 import TableSortable from './pages/examples/TableSortable/TableSortable';
@@ -6,6 +5,7 @@ import HorizontalList from './pages/examples/HorizontalList/HorizontalList';
 import StressTest from './pages/examples/StressTest';
 import FileTree from './pages/examples/FileTree/FileTree';
 import MultipleTree from './pages/examples/MultipleTree';
+import RevertOnDragOutside from './pages/examples/RevertOnDragOutside';
 
 export default [
   {
@@ -18,6 +18,7 @@ export default [
       { id: 'horizontal-list', label: 'Horizontal List', component: HorizontalList },
       { id: 'file-tree', label: 'File Tree', component: FileTree },
       { id: 'multiple-tree', label: 'Multiple Tree', component: MultipleTree },
+      { id: 'revert-on-drag-outside', label: 'Revert On Drag Outside', component: RevertOnDragOutside },
       { id: 'stress-test', label: 'Stress Test', component: StressTest },
     ]
   }