+++ packages/react-sortly/src/sortlyContext.ts
@@ -1,0 +1,10 @@
+import React from 'react';
+
+type ItemContext<D> = {
+  items: D[];
+};
+
+// @ts-ignore
+const context = React.createContext<ItemContext>({});
+
+export default context;