+++ packages/react-sortly/src/types/ItemRendererProps.ts
@@ -1,19 +1,11 @@
-import { DragElementWrapper, DragSourceOptions, DragPreviewOptions } from 'react-dnd';
-
 import ID from './ID';
 import ObjectLiteral from './ObjectLiteral';
-import ItemData from './ItemData';
 
-type ItemRendererProps<D = ObjectLiteral> = {
+type ItemRendererProps<D extends ObjectLiteral> = {
   index: number;
   id: ID;
   depth: number;
-  data: ItemData<D>;
-  drag: DragElementWrapper<DragSourceOptions>;
-  preview: DragElementWrapper<DragPreviewOptions>;
-  drop: DragElementWrapper<any>;
-  isDragging: boolean;
-  isClosestDragging: () => boolean;
+  data: D;
 };
 
 export default ItemRendererProps; // eslint-disable-line no-undef